Fix average_block_propagation so that it returns a float

+++ src/api/Monitoring.cpp
@@ -91,7 +91,7 @@ float Monitoring::average_block_propagation_since(std::time_t since) const {
   if (nb_blocks == 0) {
     return -1;
   }
-  return total_propagation / nb_blocks;
+  return static_cast<float>(total_propagation) / nb_blocks;
 }
 
 uint32_t Monitoring::nb_blocks_5m() const { return nb_blocks_since(300); }
